1st commit
"Trench Run" is a chess-like game in which each player takes turns
moving one of his/her pieces, and tries to capture the opponent's Death
Star.  It is played on an 7x7 board, and each side starts with a set of
pieces consisting of 1 Death Star, 2 walls, 4 tie fighters, and 4
x-wings. The different pieces have rules about how they can move, which
is described later.